Implement new image Provider and ImageURI token annotation
Previous abstraction change introduced GetImageURL and GetRawImage.
These methods were meant to replace GetImage, however only Discord was
using it.

Switching to these methods would require significant rewrites in
integration code, so a simpler abstraction is introduced.

This new abstraction satisfies the same requirements as the previous
in that it hides Path and File information as implementation details
so that Providers can control access to raw data.

In addition, the new GetImage no longer needs to support parsing both token and
url from annotations, as remote AM will use the new AnnotationURLProvider
instead of piggy-backing on this existing one.

That means we revert back to always using token simplifying code and security
considerations.